+++ Objects/Converters/ConverterRhinoGh/ConverterRhinoGhShared/Directory.Build.props
@@ -1,0 +1,6 @@
+<Project>
+  <!-- See https://aka.ms/dotnet/msbuild/customize for more details on customizing your build -->
+  <PropertyGroup>
+
+  </PropertyGroup>
+</Project>
