+++ Objects/Converters/ConverterRhinoGh/ConverterRhinoGhShared/Utils/IRhinoObjectSchema.cs
@@ -1,0 +1,14 @@
+using Objects.Structural.Analysis;
+using Rhino;
+using Rhino.DocObjects;
+using Speckle.Core.Kits;
+using Speckle.Core.Logging;
+using Speckle.Core.Models;
+
+namespace Objects.Converter.RhinoGh.Utils;
+
+internal interface IRhinoObjectsSchema
+{
+  string GetSchema(RhinoObject obj, string key, out string[] args);
+}
+
