+++ Objects/Converters/ConverterRhinoGh/ConverterRhinoGhShared/WIP/IRhinoObjectConverter.cs
@@ -1,0 +1,7 @@
+using Rhino.DocObjects;
+using Speckle.Core.Models;
+
+public interface IRhinoObjectConverter
+{
+  Base Convert(string appIdKey, RhinoObject ro);
+}
